requirements.txt
# AI Job Hunter Requirements
# Core dependencies for job scraping, AI analysis, and notifications

# Web scraping and automation
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0


# Telegram bot
python-telegram-bot>=20.0

# Google Sheets integration
gspread>=5.12.0
oauth2client>=4.1.3

# AI and ML
anthropic>=0.7.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
schedule>=1.2.0
pytz>=2023.3
apscheduler>=3.10.0

# Logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
